refactor(cache_control): extract cache control utilities into dedicated module
Move cache_control helper functions from chat/session.rs to a new
cache_control module for better organization and reuse:
- apply_cache_control_to_messages: marks user messages for caching
- count_system_cache_controls: counts markers in system prompt
- prune_cache_controls_in_messages: removes excess markers

Add cache_control pruning in Agent::request_params() to enforce the
API limit of 4 breakpoints by removing oldest markers when the system
prompt already uses some breakpoints.
