+++ src/lib.rs
@@ -11,6 +11,7 @@ pub mod chat;
 mod accumulating_stream;
 mod agent;
 mod backoff;
+mod cache_control;
 mod client;
 mod error;
 mod json_schema;
